+++ tests/integration/test_parallel_run.py
@@ -481,3 +481,105 @@ def test_ciceroscm_run_parallel_many_forcing(test_data_dir):
         'variable=="Surface Air Temperature Change" & scenario=="forc_data_zero_forcing" & run_id=="10496_old_NR_rounded"'
     )
     assert len(test_length.values[0]) == 358
+
+
+def test_parallel_conc_run(test_data_dir):
+    scenarios = [
+        {
+            "gaspam_file": os.path.join(test_data_dir, "gases_v1RCMIP.txt"),
+            "nyend": 2100,
+            "conc_run": True,
+            "concentrations_file": os.path.join(test_data_dir, "ssp245_conc_RCMIP.txt"),
+            "emissions_file": os.path.join(test_data_dir, "ssp245_em_RCMIP.txt"),
+            "nat_ch4_file": os.path.join(test_data_dir, "natemis_ch4.txt"),
+            "nat_n2o_file": os.path.join(test_data_dir, "natemis_n2o.txt"),
+            "sunvolc": 1,
+            "scenname": "ssp245conc",
+        }
+    ]
+
+    cfgs = [
+        {
+            "pamset_emiconc": {
+                "qh2o_ch4": 0.171,
+                "qbmb": 0.03,
+                "qo3": 0.4,
+                "qdirso2": -0.457 / 57,
+                "qindso2": -0.514 / 57,
+                "qbc": 0.200 / 7,
+                "qoc": -0.103 / 16,
+            },
+            "pamset_udm": {
+                "rlamdo": 16.0,
+                "akapa": 0.634,
+                "cpi": 0.4,
+                "W": 4.0,
+                "beto": 3.5,
+                "lambda": 0.540,
+                "mixed": 60.0,
+            },
+            "Index": "test_test",
+        },
+        {
+            "pamset_emiconc": {
+                "qh2o_ch4": 0.171,
+                "qbmb": 0.03,
+                "qo3": 0.4,
+                "qdirso2": -0.457 / 57,
+                "qindso2": -0.514 / 57,
+                "qbc": 0.200 / 7,
+                "qoc": -0.103 / 16,
+            },
+            "pamset_udm": {
+                "rlamdo": 16.0,
+                "akapa": 0.634,
+                "cpi": 0.4,
+                "W": 4.0,
+                "beto": 3.5,
+                "lambda": 0.540,
+                "mixed": 60.0,
+                "ocean_efficacy": 1.2,
+            },
+            "pamset_carbon": {"mixed_carbon": 120},
+            "Index": "test_test_oceff",
+        },
+    ]
+
+    output_vars = [
+        "Surface Air Temperature Change",
+        "Effective Radiative Forcing",
+        "Emissions|CO2",
+        "Heat Content|Ocean",
+        "Atmospheric Concentrations|CO2",
+        "Effective Radiative Forcing|CO2",
+        "Effective Radiative Forcing|N2O",
+        "Effective Radiative Forcing|CH4",
+        "Effective Radiative Forcing|Aerosols",
+        "Effective Radiative Forcing|F-Gases",
+        "Biosphere carbon flux",
+        "Airborne fraction CO2",
+        "Ocean carbon flux",
+    ]
+
+    results = run_ciceroscm_parallel(scenarios, cfgs, output_vars)
+    print(results.columns)
+    assert set(results["variable"].unique()) == set(output_vars)
+    print(results["run_id"].unique())
+    back_em1 = (
+        results.loc[
+            (results["variable"] == "Emissions|CO2")
+            & (results["run_id"] == "test_test")
+        ]
+        .iloc[:, 7:]
+        .to_numpy(float)
+    )
+    print(back_em1)
+    back_em2 = (
+        results.loc[
+            (results["variable"] == "Emissions|CO2")
+            & (results["run_id"] == "test_test_oceff")
+        ]
+        .iloc[:, 7:]
+        .to_numpy(float)
+    )
+    assert not np.allclose(back_em1, back_em2)